layer-checkboxes: Add spoofAlignment for turn/word/segment
This is useful for end-users who conceptualize alignment not as
'temporal relationship to parent' but 'temporal relationship to scope'.
While the former better reflects the annotation graph structure, the
latter is more accessible (and hence used in the APLS doc:
https://djvill.github.io/APLS/doc/layer-typology#alignment-and-horizontal-peers).
This leaves turn/word/segment as weird edge cases: they define their
scopes in a sense, but they have alignment=2 because of who their
parents are. So spoofAlignment overrides them (in display only!) to have
alignment=0.