search: If all selected participants are 'Main speaker' type, show mainParticipantOnly checkbox as checked and grayed-out
...even though it's *actually* unchecked, and the two are not strictly
equivalent (i.e., some 'Main speaker' participants may be in transcripts
in which they're not the main participant).
This is in fact the case in APLS, where (as of 0.2.2), LV09 has 76 words
in LV08pairs[12] and HD17 has 49 words in HD16interview9.
Still, I think this is worth the clarity tradeoff for end-users.